Fix serialization bugs
Worryingly, these were major bugs that stopped the program working but did not crash it or output any error messages (would be done by logging). Have investigated throwing proper exceptions to crash but due to amount of anonymous functions in use and the way checked exceptions work this is surprisingly tricky. May be best to do by handling 'failure' CommandResults in different ways depending on type of failure